perf(output): Replace Handlebars templates with direct string builders and defer heavy imports
Replace the Handlebars template engine with direct string builder functions
for XML, Markdown, and Plain output styles. This eliminates Handlebars module
loading, template compilation, and AST-based execution overhead for the main
pack pipeline.

Changes:
- Replace Handlebars templates with direct string builders (buildXmlOutput,
  buildMarkdownOutput, buildPlainOutput) using array push + join pattern
- Extract extensionToLanguageMap to fileLanguageMap.ts to break the
  Handlebars import chain from outputGenerate.ts
- Lazy-load packSkill module to prevent Handlebars transitive import
  during normal (non-skill) pack operations
- Lazy-load jschardet + iconv-lite in fileRead.ts since <1% of files
  reach the non-UTF-8 slow path
- Move metrics worker pool warmup before file search to overlap
  gpt-tokenizer initialization with I/O-bound globby traversal

Benchmark results (15 interleaved runs, repomix repo ~990 files):
- Module load time: 112ms → 133ms (old) = 57ms savings (34% reduction)
- Overall: median 2384ms → 2356ms, mean 2454ms → 2387ms
- Mean paired improvement: 67ms (2.7% on this VM)
- NEW faster in 11/15 runs

Note: VM measurement understates the improvement due to high baseline
variance. Module loading savings (~57ms) are constant regardless of
machine speed, yielding ~3.5-4% on CI baselines of 1.5-2s.
