+++ src/core/output/outputGenerate.ts
@@ -1,7 +1,6 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import XMLBuilder from 'fast-xml-builder';
-import Handlebars from 'handlebars';
 import type { RepomixConfigMerged } from '../../config/configSchema.js';
 import { RepomixError } from '../../shared/errorHandle.js';
 import { listDirectories, listFiles, searchFiles } from '../file/fileSearch.js';
@@ -9,6 +8,7 @@ import { type FilesByRoot, generateTreeString, generateTreeStringWithRoots } fro
 import type { ProcessedFile } from '../file/fileTypes.js';
 import type { GitDiffResult } from '../git/gitDiffHandle.js';
 import type { GitLogResult } from '../git/gitLogHandle.js';
+import { getLanguageFromFilePath } from './fileLanguageMap.js';
 import type { OutputGeneratorContext, RenderContext } from './outputGeneratorTypes.js';
 import { sortOutputFiles } from './outputSort.js';
 import {
@@ -19,38 +19,6 @@ import {
   generateSummaryPurpose,
   generateSummaryUsageGuidelines,
 } from './outputStyleDecorate.js';
-import { getMarkdownTemplate } from './outputStyles/markdownStyle.js';
-import { getPlainTemplate } from './outputStyles/plainStyle.js';
-import { getXmlTemplate } from './outputStyles/xmlStyle.js';
-
-// Cache for compiled Handlebars templates to avoid recompilation on every call
-const compiledTemplateCache = new Map<string, Handlebars.TemplateDelegate>();
-
-const getCompiledTemplate = (style: string): Handlebars.TemplateDelegate => {
-  const cached = compiledTemplateCache.get(style);
-  if (cached) {
-    return cached;
-  }
-
-  let template: string;
-  switch (style) {
-    case 'xml':
-      template = getXmlTemplate();
-      break;
-    case 'markdown':
-      template = getMarkdownTemplate();
-      break;
-    case 'plain':
-      template = getPlainTemplate();
-      break;
-    default:
-      throw new RepomixError(`Unsupported output style for handlebars template: ${style}`);
-  }
-
-  const compiled = Handlebars.compile(template);
-  compiledTemplateCache.set(style, compiled);
-  return compiled;
-};
 
 const calculateMarkdownDelimiter = (files: ReadonlyArray<ProcessedFile>): string => {
   const maxBackticks = files
@@ -103,6 +71,212 @@ export const createRenderContext = (outputGeneratorContext: OutputGeneratorConte
     gitLogContent: outputGeneratorContext.gitLogResult?.logContent,
     gitLogCommits: outputGeneratorContext.gitLogResult?.commits,
   };
+};
+
+// Direct string builder for XML style output (replaces Handlebars template)
+const buildXmlOutput = (ctx: RenderContext): string => {
+  const parts: string[] = [];
+
+  if (ctx.fileSummaryEnabled) {
+    parts.push(
+      `${ctx.generationHeader}\n\n<file_summary>\nThis section contains a summary of this file.\n\n<purpose>\n${ctx.summaryPurpose}\n</purpose>\n\n<file_format>\n${ctx.summaryFileFormat}\n5. Multiple file entries, each consisting of:\n  - File path as an attribute\n  - Full contents of the file\n</file_format>\n\n<usage_guidelines>\n${ctx.summaryUsageGuidelines}\n</usage_guidelines>\n\n<notes>\n${ctx.summaryNotes}\n</notes>\n\n</file_summary>\n\n`,
+    );
+  }
+
+  if (ctx.headerText) {
+    parts.push(`<user_provided_header>\n${ctx.headerText}\n</user_provided_header>\n\n`);
+  }
+
+  if (ctx.directoryStructureEnabled) {
+    parts.push(`<directory_structure>\n${ctx.treeString}\n</directory_structure>\n\n`);
+  }
+
+  if (ctx.filesEnabled) {
+    parts.push("<files>\nThis section contains the contents of the repository's files.\n");
+    for (const file of ctx.processedFiles) {
+      parts.push(`\n<file path="${file.path}">\n${file.content}\n</file>\n`);
+    }
+    parts.push('\n</files>\n');
+  }
+
+  if (ctx.gitDiffEnabled) {
+    parts.push(
+      `\n<git_diffs>\n<git_diff_work_tree>\n${ctx.gitDiffWorkTree}\n</git_diff_work_tree>\n<git_diff_staged>\n${ctx.gitDiffStaged}\n</git_diff_staged>\n</git_diffs>\n`,
+    );
+  }
+
+  if (ctx.gitLogEnabled && ctx.gitLogCommits) {
+    parts.push('\n<git_logs>');
+    for (const commit of ctx.gitLogCommits) {
+      parts.push(`\n<git_log_commit>\n<date>${commit.date}</date>\n<message>${commit.message}</message>\n<files>`);
+      for (const file of commit.files) {
+        parts.push(`\n${file}`);
+      }
+      parts.push('\n</files>\n</git_log_commit>');
+    }
+    parts.push('\n</git_logs>\n');
+  }
+
+  if (ctx.instruction) {
+    parts.push(`\n<instruction>\n${ctx.instruction}\n</instruction>`);
+  }
+
+  return parts.join('');
+};
+
+// Direct string builder for Markdown style output (replaces Handlebars template)
+const buildMarkdownOutput = (ctx: RenderContext): string => {
+  const parts: string[] = [];
+
+  if (ctx.fileSummaryEnabled) {
+    parts.push(
+      `${ctx.generationHeader}\n\n# File Summary\n\n## Purpose\n${ctx.summaryPurpose}\n\n## File Format\n${ctx.summaryFileFormat}\n5. Multiple file entries, each consisting of:\n  a. A header with the file path (## File: path/to/file)\n  b. The full contents of the file in a code block\n\n## Usage Guidelines\n${ctx.summaryUsageGuidelines}\n\n## Notes\n${ctx.summaryNotes}\n\n`,
+    );
+  }
+
+  if (ctx.headerText) {
+    parts.push(`# User Provided Header\n${ctx.headerText}\n\n`);
+  }
+
+  if (ctx.directoryStructureEnabled) {
+    parts.push(`# Directory Structure\n\`\`\`\n${ctx.treeString}\n\`\`\`\n\n`);
+  }
+
+  if (ctx.filesEnabled) {
+    const delim = ctx.markdownCodeBlockDelimiter;
+    parts.push('# Files\n\n');
+    for (const file of ctx.processedFiles) {
+      const lang = getLanguageFromFilePath(file.path);
+      parts.push(`## File: ${file.path}\n${delim}${lang}\n${file.content}\n${delim}\n\n`);
+    }
+  }
+
+  if (ctx.gitDiffEnabled) {
+    parts.push(
+      `# Git Diffs\n## Git Diffs Working Tree\n\`\`\`diff\n${ctx.gitDiffWorkTree}\n\`\`\`\n\n## Git Diffs Staged\n\`\`\`diff\n${ctx.gitDiffStaged}\n\`\`\`\n\n`,
+    );
+  }
+
+  if (ctx.gitLogEnabled && ctx.gitLogCommits) {
+    parts.push('# Git Logs\n\n');
+    for (const commit of ctx.gitLogCommits) {
+      parts.push(`## Commit: ${commit.date}\n**Message:** ${commit.message}\n\n**Files:**\n`);
+      for (const file of commit.files) {
+        parts.push(`- ${file}\n`);
+      }
+      parts.push('\n');
+    }
+  }
+
+  if (ctx.instruction) {
+    parts.push(`# Instruction\n${ctx.instruction}`);
+  }
+
+  return parts.join('');
+};
+
+const PLAIN_SEPARATOR = '='.repeat(16);
+const PLAIN_LONG_SEPARATOR = '='.repeat(64);
+
+// Direct string builder for Plain text style output (replaces Handlebars template)
+const buildPlainOutput = (ctx: RenderContext): string => {
+  const parts: string[] = [];
+
+  if (ctx.fileSummaryEnabled) {
+    parts.push(
+      `${ctx.generationHeader}\n\n${PLAIN_LONG_SEPARATOR}\nFile Summary\n${PLAIN_LONG_SEPARATOR}\n\nPurpose:\n--------\n${ctx.summaryPurpose}\n\nFile Format:\n------------\n${ctx.summaryFileFormat}\n5. Multiple file entries, each consisting of:\n  a. A separator line (================)\n  b. The file path (File: path/to/file)\n  c. Another separator line\n  d. The full contents of the file\n  e. A blank line\n\nUsage Guidelines:\n-----------------\n${ctx.summaryUsageGuidelines}\n\nNotes:\n------\n${ctx.summaryNotes}\n\n`,
+    );
+  }
+
+  if (ctx.headerText) {
+    parts.push(`\n${PLAIN_LONG_SEPARATOR}\nUser Provided Header\n${PLAIN_LONG_SEPARATOR}\n${ctx.headerText}\n\n`);
+  }
+
+  if (ctx.directoryStructureEnabled) {
+    parts.push(`${PLAIN_LONG_SEPARATOR}\nDirectory Structure\n${PLAIN_LONG_SEPARATOR}\n${ctx.treeString}\n\n`);
+  }
+
+  if (ctx.filesEnabled) {
+    parts.push(`${PLAIN_LONG_SEPARATOR}\nFiles\n${PLAIN_LONG_SEPARATOR}\n\n`);
+    for (const file of ctx.processedFiles) {
+      parts.push(`${PLAIN_SEPARATOR}\nFile: ${file.path}\n${PLAIN_SEPARATOR}\n${file.content}\n\n`);
+    }
+  }
+
+  if (ctx.gitDiffEnabled) {
+    parts.push(
+      `${PLAIN_LONG_SEPARATOR}\nGit Diffs\n${PLAIN_LONG_SEPARATOR}\n${PLAIN_SEPARATOR}\n${ctx.gitDiffWorkTree}\n${PLAIN_SEPARATOR}\n\n${PLAIN_SEPARATOR}\nGit Diffs Staged\n${PLAIN_SEPARATOR}\n${ctx.gitDiffStaged}\n\n`,
+    );
+  }
+
+  if (ctx.gitLogEnabled && ctx.gitLogCommits) {
+    parts.push(`${PLAIN_LONG_SEPARATOR}\nGit Logs\n${PLAIN_LONG_SEPARATOR}\n`);
+    for (const commit of ctx.gitLogCommits) {
+      parts.push(`${PLAIN_SEPARATOR}\nDate: ${commit.date}\nMessage: ${commit.message}\nFiles:\n`);
+      for (const file of commit.files) {
+        parts.push(`  - ${file}\n`);
+      }
+      parts.push(`${PLAIN_SEPARATOR}\n\n`);
+    }
+    parts.push('\n');
+  }
+
+  if (ctx.instruction) {
+    parts.push(`${PLAIN_LONG_SEPARATOR}\nInstruction\n${PLAIN_LONG_SEPARATOR}\n${ctx.instruction}`);
+  }
+
+  parts.push(`\n${PLAIN_LONG_SEPARATOR}\nEnd of Codebase\n${PLAIN_LONG_SEPARATOR}\n`);
+
+  return parts.join('');
+};
+
+const generateDirectOutput = async (
+  config: RepomixConfigMerged,
+  renderContext: RenderContext,
+  processedFiles?: ProcessedFile[],
+): Promise<string> => {
+  try {
+    let output: string;
+    switch (config.output.style) {
+      case 'xml':
+        output = buildXmlOutput(renderContext);
+        break;
+      case 'markdown':
+        output = buildMarkdownOutput(renderContext);
+        break;
+      case 'plain':
+        output = buildPlainOutput(renderContext);
+        break;
+      default:
+        throw new RepomixError(`Unsupported output style: ${config.output.style}`);
+    }
+    return `${output.trim()}\n`;
+  } catch (error) {
+    if (error instanceof RangeError && error.message === 'Invalid string length') {
+      let largeFilesInfo = '';
+      if (processedFiles && processedFiles.length > 0) {
+        const topFiles = processedFiles
+          .sort((a, b) => b.content.length - a.content.length)
+          .slice(0, 5)
+          .map((f) => `  - ${f.path} (${(f.content.length / 1024 / 1024).toFixed(1)} MB)`)
+          .join('\n');
+        largeFilesInfo = `\n\nLargest files in this repository:\n${topFiles}`;
+      }
+
+      throw new RepomixError(
+        `Output size exceeds JavaScript string limit. The repository contains files that are too large to process.
+Please try:
+  - Use --ignore to exclude large files (e.g., --ignore "docs/**" or --ignore "*.html")
+  - Use --include to process only specific files
+  - Process smaller portions of the repository at a time${largeFilesInfo}`,
+        { cause: error },
+      );
+    }
+    throw new RepomixError(
+      `Failed to generate output: ${error instanceof Error ? error.message : 'Unknown error'}`,
+      error instanceof Error ? { cause: error } : undefined,
+    );
+  }
 };
 
 const generateParsableXmlOutput = async (renderContext: RenderContext): Promise<string> => {
@@ -214,42 +388,6 @@ const generateParsableJsonOutput = async (renderContext: RenderContext): Promise
   }
 };
 
-const generateHandlebarOutput = async (
-  config: RepomixConfigMerged,
-  renderContext: RenderContext,
-  processedFiles?: ProcessedFile[],
-): Promise<string> => {
-  try {
-    const compiledTemplate = getCompiledTemplate(config.output.style);
-    return `${compiledTemplate(renderContext).trim()}\n`;
-  } catch (error) {
-    if (error instanceof RangeError && error.message === 'Invalid string length') {
-      let largeFilesInfo = '';
-      if (processedFiles && processedFiles.length > 0) {
-        const topFiles = processedFiles
-          .sort((a, b) => b.content.length - a.content.length)
-          .slice(0, 5)
-          .map((f) => `  - ${f.path} (${(f.content.length / 1024 / 1024).toFixed(1)} MB)`)
-          .join('\n');
-        largeFilesInfo = `\n\nLargest files in this repository:\n${topFiles}`;
-      }
-
-      throw new RepomixError(
-        `Output size exceeds JavaScript string limit. The repository contains files that are too large to process.
-Please try:
-  - Use --ignore to exclude large files (e.g., --ignore "docs/**" or --ignore "*.html")
-  - Use --include to process only specific files
-  - Process smaller portions of the repository at a time${largeFilesInfo}`,
-        { cause: error },
-      );
-    }
-    throw new RepomixError(
-      `Failed to compile template: ${error instanceof Error ? error.message : 'Unknown error'}`,
-      error instanceof Error ? { cause: error } : undefined,
-    );
-  }
-};
-
 export const generateOutput = async (
   rootDirs: string[],
   config: RepomixConfigMerged,
@@ -260,7 +398,7 @@ export const generateOutput = async (
   filePathsByRoot?: FilesByRoot[],
   deps = {
     buildOutputGeneratorContext,
-    generateHandlebarOutput,
+    generateDirectOutput,
     generateParsableXmlOutput,
     generateParsableJsonOutput,
     sortOutputFiles,
@@ -287,12 +425,12 @@ export const generateOutput = async (
     case 'xml':
       return config.output.parsableStyle
         ? deps.generateParsableXmlOutput(renderContext)
-        : deps.generateHandlebarOutput(config, renderContext, sortedProcessedFiles);
+        : deps.generateDirectOutput(config, renderContext, sortedProcessedFiles);
     case 'json':
       return deps.generateParsableJsonOutput(renderContext);
     case 'markdown':
     case 'plain':
-      return deps.generateHandlebarOutput(config, renderContext, sortedProcessedFiles);
+      return deps.generateDirectOutput(config, renderContext, sortedProcessedFiles);
     default:
       throw new RepomixError(`Unsupported output style: ${config.output.style}`);
   }
