+++ tests/core/output/outputGenerate.test.ts
@@ -17,7 +17,7 @@ const createStrictXmlParser = () => {
 describe('outputGenerate', () => {
   const mockDeps = {
     buildOutputGeneratorContext: vi.fn(),
-    generateHandlebarOutput: vi.fn(),
+    generateDirectOutput: vi.fn(),
     generateParsableXmlOutput: vi.fn(),
     generateParsableJsonOutput: vi.fn(),
     sortOutputFiles: vi.fn(),
@@ -48,7 +48,7 @@ describe('outputGenerate', () => {
       instruction: '',
       filesEnabled: true,
     });
-    mockDeps.generateHandlebarOutput.mockResolvedValue('mock output');
+    mockDeps.generateDirectOutput.mockResolvedValue('mock output');
 
     const output = await generateOutput(
       [process.cwd()],
