+++ tests/core/output/outputGenerateDiffs.test.ts
@@ -59,7 +59,7 @@ describe('Output Generation with Diffs', () => {
       instruction: '',
       gitDiffResult,
     })),
-    generateHandlebarOutput: vi.fn(),
+    generateDirectOutput: vi.fn(),
     generateParsableXmlOutput: vi.fn(),
     generateParsableJsonOutput: vi.fn(),
     sortOutputFiles: vi.fn().mockResolvedValue(mockProcessedFiles),
@@ -71,7 +71,7 @@ describe('Output Generation with Diffs', () => {
     mockConfig.output.parsableStyle = false;
 
     // Mock the Handlebars output function to check for diffs in the template
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    mockDeps.generateDirectOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
       // Verify that the renderContext has the gitDiffs property
       expect(renderContext.gitDiffWorkTree).toBe(sampleDiff);
 
@@ -96,8 +96,8 @@ describe('Output Generation with Diffs', () => {
     expect(output).toContain(sampleDiff);
     expect(output).toContain('</diffs>');
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the generateDirectOutput function was called
+    expect(mockDeps.generateDirectOutput).toHaveBeenCalled();
   });
 
   test('XML style output with parsableStyle should include diffs section', async () => {
@@ -141,7 +141,7 @@ describe('Output Generation with Diffs', () => {
     mockConfig.output.parsableStyle = false;
 
     // Mock the Handlebars output function for markdown
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    mockDeps.generateDirectOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
       // Verify that the renderContext has the gitDiffs property
       expect(renderContext.gitDiffWorkTree).toBe(sampleDiff);
 
@@ -167,8 +167,8 @@ describe('Output Generation with Diffs', () => {
     expect(output).toContain(sampleDiff);
     expect(output).toContain('```');
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the generateDirectOutput function was called
+    expect(mockDeps.generateDirectOutput).toHaveBeenCalled();
   });
 
   test('Plain style output should include diffs section when includeDiffs is enabled', async () => {
@@ -177,7 +177,7 @@ describe('Output Generation with Diffs', () => {
     mockConfig.output.parsableStyle = false;
 
     // Mock the Handlebars output function for plain text
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    mockDeps.generateDirectOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
       expect(renderContext.gitDiffWorkTree).toBe(sampleDiff);
 
       // Simulate the plain text output
@@ -200,8 +200,8 @@ describe('Output Generation with Diffs', () => {
     expect(output).toContain('===============\nGit Diffs\n===============');
     expect(output).toContain(sampleDiff);
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the generateDirectOutput function was called
+    expect(mockDeps.generateDirectOutput).toHaveBeenCalled();
   });
 
   test('Output should not include diffs section when includeDiffs is disabled', async () => {
@@ -221,7 +221,7 @@ describe('Output Generation with Diffs', () => {
     }));
 
     // Mock the Handlebars output function
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    mockDeps.generateDirectOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
       // Verify that the renderContext does not have the gitDiffs property
       expect(renderContext.gitDiffWorkTree).toBeUndefined();
 
@@ -245,7 +245,7 @@ describe('Output Generation with Diffs', () => {
     expect(output).not.toContain('Git Diffs');
     expect(output).not.toContain(sampleDiff);
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the generateDirectOutput function was called
+    expect(mockDeps.generateDirectOutput).toHaveBeenCalled();
   });
 });
